+++ app/src/main/java/lol/terabrendon/houseshare2/data/remote/dto/ShoppingItemDto.kt
@@ -9,7 +9,7 @@ data class ShoppingItemDto(
     val groupId: Long,
     val name: String,
     val amount: Int,
-    val price: Double?,
+    val price: Long?,
     val priority: ShoppingItemPriority,
     val createdAt: OffsetDateTime,
     val check: CheckDto?,
